update angular version

+++ src/app/components/note-card/note-card.component.ts
@@ -1,26 +1,24 @@
-import {Component, Input, OnInit} from '@angular/core';
+import { Component, Input, OnInit } from '@angular/core';
 import { Note } from 'src/app/models/note';
 import { NotesService } from 'src/app/services/notes-service';
 
-declare const feather;
+declare const feather: any;
 
 @Component({
   selector: 'subject-note-card',
   templateUrl: './note-card.component.html',
-  styleUrls: ['./note-card.component.scss']
+  styleUrls: ['./note-card.component.css'],
 })
 export class NoteCardComponent implements OnInit {
-  @Input() note: Note;
+  @Input() note?: Note;
 
-  constructor(private notesService: NotesService) {
-  }
+  constructor(private notesService: NotesService) {}
 
   ngOnInit() {
     feather.replace();
   }
 
   deleteNote() {
-    this.notesService.deleteNote(this.note.id)
+    if (this.note) this.notesService.deleteNote(this.note.id);
   }
-
 }
